Split up interfaces.d.ts and move it to src/types. Also export constants in package.json.
