+++ tsconfig.json
@@ -13,5 +13,6 @@
     "esModuleInterop": true,
     "moduleResolution": "node16"
   },
-  "include": ["src"]
+  "include": ["src"],
+  "exclude": ["src/bundle.ts"]
 }
